ch5/Figure_5.3_5.9/requirements.txt
scipy
numpy
matplotlib
pytorch
